Implemented an abstract class to define circles.

+++ Mini Golf/src/in/vvest/server/Packet.java
@@ -103,24 +103,24 @@ public class Packet {
 			addDouble(ch.getPos().x);
 			addDouble(ch.getPos().y);
 			addDouble(ch.getRadius());
-			addDouble(ch.getStart());
-			addDouble(ch.getEnd());
-			addDouble(ch.getAcc());			
+			addDouble((double) ch.getStart());
+			addDouble((double) ch.getEnd());
+			addDouble((double) ch.getAcc());			
 		} else if (o.getID() == ObstacleID.CIRCLE_LINE) {
 			CircleLine cl = (CircleLine) o;
 			addDouble(cl.getPos().x);
 			addDouble(cl.getPos().y);
 			addDouble(cl.getRadius());
-			addDouble(cl.getStart());
-			addDouble(cl.getEnd());
+			addDouble((double) cl.getStart());
+			addDouble((double) cl.getEnd());
 			addBoolean(cl.isFilled());
 		} else if (o.getID() == ObstacleID.CIRCLE_WALL) {
 			CircleWall cw = (CircleWall) o;
 			addDouble(cw.getPos().x);
 			addDouble(cw.getPos().y);
 			addDouble(cw.getRadius());
-			addDouble(cw.getStart());
-			addDouble(cw.getEnd());
+			addDouble((double) cw.getStart());
+			addDouble((double) cw.getEnd());
 			addDouble(cw.getThickness());
 		} else if (o.getID() == ObstacleID.HOLE) {
 			Hole h = (Hole) o;
